# packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: KULTIVAR API
  version: 1.0.0
  description: |
    REST API for the KULTIVAR habit tracker.

    **Auth:** HttpOnly cookies (`accessToken`, `refreshToken`). Use `credentials: "include"` on all requests.

    **Status:** Auth, Profile, Goals, Life Areas, Habits, and Check-ins are implemented.
servers:
  - url: /api
    description: Default server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: HttpOnly access token cookie set on login
...
  schemas:
...
    Frequency:
      type: string
      enum: [daily, weekdays, 3x, 5x, weekly]
    Difficulty:
      type: string
      enum: [easy, medium, hard]
    Habit:
      type: object
      properties:
        id:
          type: string
          format: uuid
        areaId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        frequency:
          $ref: "#/components/schemas/Frequency"
        durationMinutes:
          type: integer
          minimum: 1
          maximum: 1440
          nullable: true
        difficulty:
          allOf:
            - $ref: "#/components/schemas/Difficulty"
          nullable: true
        notes:
          type: string
          maxLength: 500
          nullable: true
        reminderEnabled:
          type: boolean
          description: Whether a reminder is scheduled for this habit.
        reminderTime:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          example: "07:30"
          nullable: true
          description: 24-hour local time the reminder fires at. Required when reminderEnabled is true.
        timezone:
          type: string
          example: "America/New_York"
          nullable: true
          description: IANA timezone reminderTime is expressed in. Required when reminderEnabled is true.
        daysOfWeek:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
          nullable: true
          example: [1, 3, 5]
          description: >-
            Explicit weekdays (0=Sun..6=Sat) this habit runs on. Only
            meaningful for "3x"/"5x"/"weekly" frequencies — "daily" and
            "weekdays" already imply their own days. Null means unspecified,
            in which case reminders fire every day instead.
        archivedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CoachUpdateHabitBody:
      type: object
      description: >-
        The shape of the plan, and nothing else. Reminder settings and the
        habit's life area are not editable by a coach — those are the client's
        own; see the schema source for the reasoning.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        frequency:
          $ref: "#/components/schemas/Frequency"
        daysOfWeek:
          type: array
          nullable: true
          minItems: 1
          maxItems: 7
          items:
            type: integer
            minimum: 0
            maximum: 6
        durationMinutes:
          type: integer
          nullable: true
          minimum: 1
          maximum: 1440
        difficulty:
          type: string
          nullable: true
          enum: [easy, medium, hard]
        notes:
          type: string
          nullable: true
          maxLength: 500
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid credentials
...
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Invalid email address
paths:
...
  /coach-requests/{id}/habits/{habitId}:
    patch:
      tags: [Coaching]
      summary: Adjust a client's habit (coach only, with permission)
      operationId: updateClientHabit
      description: >-
        Requires the client's `editHabits` grant; 403 without it, and the
        grant can be withdrawn at any moment. Every change that alters
        something is written into the feedback thread as a `habit_change`
        entry, so the client is never surprised by an edited plan.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The coaching request id
          schema:
            type: string
            format: uuid
        - name: habitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachUpdateHabitBody"
      responses:
        "200":
          description: The updated habit
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Habit"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not this request's coach, not accepted, or no edit permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Request or habit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"